Strategy#page_headers: Use options argument
This approach is used in the forthcoming `options` feature as a way
of passing in options that control behavior. For the moment, this will
allow us to pass in additional curl args, so we can modify
`HeaderMatch` to retry a request using `GET` when the `HEAD` request
doesn't successfully provide version information.