Strategy#page_headers: Use `curl_args` option
This implements support for a `:curl_args` option, where this array
is simply appended to the default `PAGE_HEADERS_CURL_ARGS`. This is
fine with respect to our current setup but theoretically we could
encounter a situation in the future where we need to _unset_ some of
the default curl args and that would require a different solution.