HeaderMatch: Retry with a GET request
This implements temporary logic in the `HeaderMatch` strategy to
retry a check using a `GET` request when it fails or doesn't produce
version information.

In the near future, the `options` feature will allow us to specify
additional curl args in a `livecheck` block, so we will be able to
remove this temporary logic and account for this in the specific
checks where using a `GET` request is necessary. That approach is
preferable, as this retry logic wastes a `HEAD` request before trying
`GET` and can also end up retrying requests where using `GET` will
also fail. Ideally, we want to avoid making unnecessary requests to a
server when possible, so this is only intended as a short-term
workaround.