Strategy: Omit unnecessary curl options
`#page_headers` and `#page_content` are both now using `#curl_output`
either directly or indirectly, so the default values for the
`print_stdout` and `print_stderr` options should align with the
existing values in `DEFAULT_CURL_OPTIONS` and can now be omitted.

I ran some tests with `verbose` removed and it doesn't *appear* to
have any visible effect. I can't remember if I originally set it in
response to a situation that we experience or if it was a precaution
(after experiencing issues with `debug`) but I think it was only the
latter. If this removal causes issues, we can always add it back.