GitHub: Allow passing curl options to some methods
This commit makes it possible to pass additional curl options to
`GitHub#get_latest_release` and `GitHub::API#open_rest`. These
options are passed on to `#curl_output` and `#curl_args` in turn.
This is only the subset of `#curl_with_workarounds` and `#curl_args`
options that `livecheck` currently uses.

This is intended to ensure that curl behaves in a way that's
appropriate for livecheck. For example, `brew livecheck --debug`
output is carefully crafted but `Utils::Curl` methods print
additional debug output in the middle unless we override it with
`debug: false`. In this scenario, we need to be able to pass
`debug: false` to `#curl_options` to prevent its debug output from
printing and that's one use case that this supports.